+++ vcd/provider.go
@@ -64,6 +64,9 @@ var globalDataSourceMap = map[string]*schema.Resource{
 	"vcd_vcenter":             datasourceVcdVcenter(),           // 3.0
 	"vcd_resource_list":       datasourceVcdResourceList(),      // 3.1
 	"vcd_resource_schema":     datasourceVcdResourceSchema(),    // 3.1
+	"vcd_nsxt_edge_cluster":   datasourceVcdNsxtEdgeCluster(),   // 3.1
+	"vcd_nsxt_edgegateway":    datasourceVcdNsxtEdgeGateway(),   // 3.1
+	"vcd_storage_profile":     datasourceVcdStorageProfile(),    // 3.1
 }
 
 var globalResourceMap = map[string]*schema.Resource{
@@ -106,6 +109,7 @@ var globalResourceMap = map[string]*schema.Resource{
 	"vcd_vapp_access_control":  resourceVcdAccessControlVapp(),        // 3.0
 	"vcd_external_network_v2":  resourceVcdExternalNetworkV2(),        // 3.0
 	"vcd_vm_sizing_policy":     resourceVcdVmSizingPolicy(),           // 3.0
+	"vcd_nsxt_edgegateway":     resourceVcdNsxtEdgeGateway(),          // 3.1
 }
 
 // Provider returns a terraform.ResourceProvider.